:recycle: move `find` function from `__init__.py` to `public.py`
