+++ discord/utils/__init__.py
@@ -49,6 +49,7 @@ from .public import (
     basic_autocomplete,
     generate_snowflake,
     utcnow,
+    find,
     snowflake_time,
     oauth_url,
     Undefined,
@@ -208,32 +209,6 @@ def copy_doc(original: Callable) -> Callable[[T], T]:
         return overridden
 
     return decorator
-
-
-def find(predicate: Callable[[T], Any], seq: Iterable[T]) -> T | None:
-    """A helper to return the first element found in the sequence
-    that meets the predicate. For example: ::
-
-        member = discord.utils.find(lambda m: m.name == "Mighty", channel.guild.members)
-
-    would find the first :class:`~discord.Member` whose name is 'Mighty' and return it.
-    If an entry is not found, then ``None`` is returned.
-
-    This is different from :func:`py:filter` due to the fact it stops the moment it finds
-    a valid entry.
-
-    Parameters
-    ----------
-    predicate
-        A function that returns a boolean-like result.
-    seq: :class:`collections.abc.Iterable`
-        The iterable to search through.
-    """
-
-    for element in seq:
-        if predicate(element):
-            return element
-    return None
 
 
 async def get_or_fetch(obj, attr: str, id: int, *, default: Any = MISSING) -> Any:
